moved: deferred.ts -> util
