+++ test.ts
@@ -2,7 +2,7 @@
 // Copyright 2018-2019 the Deno authors. All rights reserved. MIT license.
 
 import "benching/test.ts";
-import "async/deferred_test.ts";
+import "util/deferred_test.ts";
 import "colors/test.ts";
 import "datetime/test.ts";
 import "examples/test.ts";
@@ -17,7 +17,6 @@ import "io/test.ts";
 import "http/server_test.ts";
 import "http/file_server_test.ts";
 import "http/readers_test.ts";
-import "http/path_to_regexp_test.ts";
 import "log/test.ts";
 import "media_types/test.ts";
 import "multipart/formfile_test.ts";
